add a README for the shared library
