+++ cmd/shared/.gitignore
@@ -1,3 +1,3 @@
 lib.a
 lib.h
-main
+example
